authen-service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
loguru
python-jose[cryptography]